+++ R/core-latex.R
@@ -73,3 +73,18 @@ underline_off_latex <- "}"
 # })"
 
 
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# Escape latex
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+escape_latex <- function (x, newlines = FALSE, spaces = TRUE) {
+  x = gsub("\\\\", "\\\\textbackslash", x)
+  x = gsub("([#$%&_{}])", "\\\\\\1", x)
+  x = gsub("\\\\textbackslash", "\\\\textbackslash{}", x)
+  x = gsub("~", "\\\\textasciitilde{}", x)
+  x = gsub("\\^", "\\\\textasciicircum{}", x)
+  if (newlines)
+    x = gsub("(?<!\n)\n(?!\n)", "\\\\\\\\", x, perl = TRUE)
+  if (spaces)
+    x = gsub("(?<= ) ", "\\\\ ", x, perl = TRUE)
+  x
+}