+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# emphatic 0.1.6.9005  2024-05-01
+# emphatic 0.1.6.9006  2024-05-01
 
 * Removed `hl_mat()` and associated vignettes.
     * function is not exported for now as the selection process is too 
@@ -11,6 +11,7 @@
   rendered to PDF
 * Remove `scale_mode` argument to `hl()`
 * Support for output to `typst` in quarto docs
+    * Updated the list of characters to escape for typst
 
 # emphatic 0.1.6  2024-04-27